refactor: reorganize whole frontend/src file structure
