+++ frontend/src/atoms/EditableField.tsx
@@ -5,18 +5,18 @@ import {
   FormMessage,
 } from "@/lib/shadcn/generated/ui/form";
 import type { ControllerProps, FieldPath, FieldValues } from "react-hook-form";
-import { Input } from "../atoms/Input";
+import { Input } from "./Input";
 
 type Props<
   TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
+  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>
 > = Pick<ControllerProps<TFieldValues, TName>, "control" | "name"> & {
   label: string;
 };
 
 export function EditableField<
   TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
+  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>
 >({ label, name, control }: Props<TFieldValues, TName>) {
   return (
     <FormField
