+++ frontend/src/layout/AppSidebar.tsx
@@ -17,8 +17,8 @@ import {
   SidebarHeader,
 } from "@/lib/shadcn/generated/ui/sidebar";
 import { Link } from "react-router-dom";
-import CampaignForm from "./CampaignForm";
-import ScenarioForm from "./ScenarioForm";
+import CampaignForm from "../organisms/Campaign/CampaignForm";
+import ScenarioForm from "../organisms/Scenario/ScenarioForm";
 
 export default function AppSidebar() {
   return (
