+++ frontend/src/organisms/Flashcard/FlashCard.tsx
