+++ frontend/src/pages/public/ErrorPage.tsx
